// Front-Angular/BibliotecaOnline/src/app/libros/interfaces/libro.interface.ts
export interface Libro {
    id:          number;
    title:       string;
    date:        string;
    author:      string;
    available:   number;
    stock:       number;
    category:    string;
    description: string;
    lang:        string;
    pages:        number;
    edit:        string;
    img:         string;
    copies:      string;
}